feat(initialState): return initial state when creating the store
initial state is only available once a subscription is made to the state$. In React it's recommended
to use componentDidMount instead of componentWillMount. Setting state in componentDidMount is too
late at this point as the component has already attempted to render with a state of 'null'. By
returning the initialState of the store we can provide components with the initialState before a
subscription is created, eliminating any need for null checks on objects